Add ModalView for routing modals
