+++ client/src/views/UserSigninModal.vue
@@ -1,27 +1,16 @@
 <script setup>
 import useAuthStore from '@/stores/auth'
-import { ref, computed } from 'vue'
-import { useRoute, useRouter } from 'vue-router'
+import { ref } from 'vue'
+import { useRoute } from 'vue-router'
 
 const route = useRoute()
-const router = useRouter()
 const auth = useAuthStore()
 const form = ref({})
 const { execute, error, data } = auth.signIn(form, { immediate: false })
-
-const isOpen = computed(() => route.hash === '#signin' && !auth.isAuthenticated)
-function dismissed () {
-  if (isOpen.value) router.back()
-}
 </script>
 
 <template>
-  <ion-modal
-    :is-open="isOpen"
-    can-dismiss
-    :presenting-element="$parent.$refs.ionRouterOutlet"
-    @did-dismiss="dismissed()"
-  >
+  <ion-modal>
     <ion-page>
       <Head>
         <title>Sign In</title>
